Added smoke test file
